% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{check_version_nf}
\alias{check_version_nf}
\title{Check which version(s) of the NF are present in a data set.}
\usage{
check_version_nf(dat)
}
\arguments{
\item{dat}{a data.frame}
}
\value{
A character vector containing one or more of the stable version
identifiers \code{"2"} and \code{"3"}, in that order. If no matching items are found,
returns \code{"No NF items found."}. Raw export names are supported when they
contain an item token such as \code{M1172_Q142}.
}
\description{
Should not require the data to already have clean names. But needs to work when that is the case.
}
\examples{
\dontrun{
check_version_nf(dataGoesHere)
}
}
