% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{find_trigger}
\alias{find_trigger}
\title{Identify the trigger item for a named scale}
\usage{
find_trigger(scaleName)
}
\arguments{
\item{scaleName}{The name of the scale to look up. Must be one of the NF2 coding names}
}
\value{
item name of the trigger item, e.g., "Q51"
}
\description{
Identify the trigger item for a named scale
}
\examples{
suppressWarnings(find_trigger("somAnx"))  # "Q51"

# Can be used in concert with get_item_text():
get_item_text(suppressWarnings(find_trigger("somAnx")))
# "I have a sense of restlessness and unease in me most of the time"
}
