% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{find_trigger_among}
\alias{find_trigger_among}
\title{Identify which of the items is a trigger item}
\usage{
find_trigger_among(items)
}
\arguments{
\item{items}{A vector of quoted item names, e.g., somAnx.names}
}
\value{
The item name that is identified as a trigger in \code{item_descriptions}.
}
\description{
Identify which of the items is a trigger item
}
\examples{
find_trigger_among(somAnx.names)
}
