% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningFuncs.R
\name{get_first_obs}
\alias{get_first_obs}
\title{Pick out only the first row per patient of a NF-type data frame.}
\usage{
get_first_obs(.data, id = "respondent id")
}
\arguments{
\item{.data}{Data set on which to select rows}

\item{id}{Name of variable representing patient id. Default is \verb{respondent id}.
Works if variable name is quoted or not.}
}
\value{
A tibble.
}
\description{
Note: This function does not sort the data by date.
}
\examples{
data(synthetic_data, package = "NorseResearch")
get_first_obs(synthetic_data, id = "anon_id")
}
