% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningFuncs.R
\name{get_nicer_name}
\alias{get_nicer_name}
\title{Swap the short name of a scale for the nicer name of a scale.}
\usage{
get_nicer_name(simplename, version = "2")
}
\arguments{
\item{simplename}{A character matching a value in \code{scale_names}.}

\item{version}{A character string defining version to use. Defaults to
\code{2} for NF 2.x.}
}
\value{
A character from \code{nicer.nf2.names}. If no match found, returns \code{simplename}.
}
\description{
Swap the short name of a scale for the nicer name of a scale.
}
\examples{
get_nicer_name("eating") # should return "Eating Problems"
get_nicer_name(c("eating", "subUse")) # should return a character vector
}
