% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R
\name{iic.grm}
\alias{iic.grm}
\title{Plotting Item Information Curves}
\usage{
iic.grm(x, lim = c(-4, 4))
}
\arguments{
\item{x}{fitted object.}

\item{lim}{limits for x-axis. Defaults to c(-4, 4).}
}
\description{
\code{icc.grm} takes a fitted ltm model and plots its Item information curves
}
