% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R
\name{info_mirt}
\alias{info_mirt}
\title{Marginal information per item. Only works with {mirt} objects.}
\usage{
info_mirt(
  fit,
  z = c(-6, 6),
  n.items = length(dimnames(fit@Data$data)[[2]]),
  printAuto = TRUE
)
}
\arguments{
\item{fit}{an IRT fitted model, built for those derived from package \code{mirt}.}

\item{z}{a vector, upper and lower limits on theta of interest. Default is c(-6,6)}

\item{n.items}{the number of items in the scale, defaults to the full scale from \code{fit}.}

\item{printAuto}{Should the info table be printed? Defaults to TRUE for direct calls.}
}
\value{
a table with per-item marginal information over the range supplied by \code{z}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{info_mirt} provides a quick summary of item performance in terms of marginal
information over some range of theta. Use \code{\link[=info]{info()}} instead.
}
\examples{
info_mirt(cogGP_mirt)
}
