% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{internal_data}
\alias{internal_data}
\alias{alliance.names}
\alias{alliance.names.nf3}
\alias{anger.names.nf3}
\alias{avoidSit.names}
\alias{avoidSoc.names}
\alias{cog.names}
\alias{cog.names.nf3}
\alias{control.names}
\alias{eating.names}
\alias{eating.names.nf3}
\alias{genFunc.names}
\alias{genFunc.names.nf3}
\alias{hopeless.names}
\alias{hopeless.names.nf3}
\alias{impulsivity.names.nf3}
\alias{intAvoid.names.nf3}
\alias{internal.names}
\alias{internal.names.nf3}
\alias{intMem.names.nf3}
\alias{irritable.names}
\alias{list_objs}
\alias{names.list}
\alias{needs.names}
\alias{needs.names.nf3}
\alias{nicer_names_nf3}
\alias{nicer.nf2.names}
\alias{ona.names}
\alias{ona.names.nf3}
\alias{pain.names.nf3}
\alias{physAnx.names.nf3}
\alias{pref.names.nf3}
\alias{ready.names}
\alias{ready.names.nf3}
\alias{recovEnv.names}
\alias{sad.names}
\alias{sad.names.nf3}
\alias{scale_names}
\alias{scale_names_nf3}
\alias{scale_names_ou}
\alias{SDH.names.nf3}
\alias{selfComp.names.nf3}
\alias{selfContempt.names.nf3}
\alias{selfCrit.names}
\alias{single.items.names}
\alias{single.items.names.nf3}
\alias{socAvoid.names.nf3}
\alias{socSup.names.nf3}
\alias{socialSafety.names}
\alias{somAnx.names}
\alias{subRecov.names}
\alias{subUse.names}
\alias{subUse.names.nf3}
\alias{suicide.names}
\alias{suicide.names.nf3}
\alias{trauma.names}
\alias{worry.names}
\alias{worry.names.nf3}
\title{Internal scoring and lookup data}
\description{
Item-name vectors, scale-name mappings, and compatibility objects used
internally by NorseResearch. They are retained as package data for
backwards compatibility but are not generally needed directly.
}
\keyword{internal}
