% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{item_names_nf2}
\alias{item_names_nf2}
\title{Names of items}
\format{
An character vector of item names.
}
\source{
See \code{data-raw/item_names_nf2.R}
}
\usage{
item_names_nf2
}
\description{
Names of items
}
\keyword{datasets}
