% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFuncs.R
\name{lookup_item}
\alias{lookup_item}
\title{Look up information about an item}
\usage{
lookup_item(target, version = "2.1", verbose = FALSE)
}
\arguments{
\item{target}{Item quoted as character (e.g., "Q61")}

\item{version}{Norse version to use. Changes the lookup table from 'item_descriptions'
(used for v. 2.0) to 'nf2.1.item.descriptions' (used for 2.1).}

\item{verbose}{Logical. Provide more than just item text?}
}
\value{
If \code{verbose = FALSE}, a string of item text.
If \code{verbose = TRUE}, a single row data frame.
}
\description{
Look up information about an item
}
\examples{
lookup_item("Q145")
# no difference:
lookup_item("Q145", version = 2)

# version can be any of the following: 2, 2.0, "2", 2.1, or "2.1".

# compare v 2 to 2.1 using verbose = TRUE
lookup_item("Q61", version = 2, verbose = TRUE)    # It was the trigger in v2.0
lookup_item("Q61", version = 2.1, verbose = TRUE)  # but not in v2.1
}
\seealso{
\code{\link{get_item_text}}, a deprecated version but necessary for re-running analyses.
}
