% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFuncs.R
\name{lookup_trigger}
\alias{lookup_trigger}
\title{Look up the trigger item given a scale name in either programming or presentation format.}
\usage{
lookup_trigger(scaleName, version = 2.1)
}
\arguments{
\item{scaleName}{A (partial) scale name or simple name. "Somatic Anxiety",
"somAnx", and "Somatic" should all work for version = 2.1.}

\item{version}{Version of the NF to use. Changes the lookup table from 'item_descriptions'
(used for v. 2.0) to 'nf2.1.logic' (used for 2.1).}
}
\value{
A character item name of the trigger.
}
\description{
Look up the trigger item given a scale name in either programming or presentation format.
}
\examples{
lookup_trigger("hopeless")
# "Q115"
lookup_trigger("hopeless", version = 2)
# "Q61"

# works with nicer names in v. 2.1 ONLY:
lookup_trigger("Hopeless")
# "Q115"

# when lookup fails:
\dontrun{
lookup_trigger("Hopless")  # easy misspelling
}
# "scaleName not recognized"
# NULL

}
