% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFuncs.R
\name{lookup_trigger_among}
\alias{lookup_trigger_among}
\title{Look up trigger item among a list of item names}
\usage{
lookup_trigger_among(items, version = 2.1)
}
\arguments{
\item{items}{A vector or list of quoted item names,
e.g., somAnx.names}

\item{version}{Version of the NF to use. Changes the lookup table from 'item_descriptions'
(used for v. 2.0) to 'nf2.1.logic' (used for 2.1).}
}
\value{
The item name that is identified as a trigger in the appropriate version.
If none of the items are identified as a trigger, returns \code{NA} with a warning.
When multiple trigger items are detected, returns a character vector with a message.
}
\description{
Look up trigger item among a list of item names
}
\examples{
lookup_trigger_among(somAnx.names)

# sends a message:
lookup_trigger_among(append(somAnx.names, sad.names))

# returns NA with warning:
lookup_trigger_among(subRecov.names)
}
