% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{random_item_response}
\alias{random_item_response}
\title{Item response RNG}
\usage{
random_item_response()
}
\value{
A single number, 1:7
}
\description{
Item response RNG
}
\examples{
random_item_response()
}
