% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{random_item_vector}
\alias{random_item_vector}
\title{Random item responses as a vector}
\usage{
random_item_vector(num = 1, range = 1:7, replace = TRUE)
}
\arguments{
\item{num}{Numeric. How many items?}

\item{range}{Vector. Defaults to 1:7}

\item{replace}{Logical. Should observations be repeatable? Defaults to TRUE.}
}
\value{
a vector of length \code{num}
}
\description{
Random item responses as a vector
}
\examples{
random_item_vector(20)
}
