% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{random_person_generator}
\alias{random_person_generator}
\title{Generate a set of random patient variables of arbitrary length}
\usage{
random_person_generator(
  num = 2,
  mean_obs = 5.79,
  gender = NULL,
  birthyear = NULL,
  tx_focus = NULL,
  in_or_out = NULL,
  ...
)
}
\arguments{
\item{num}{Integer. Number of people.}

\item{mean_obs}{Numeric. Defines a typical number of observations per person. Is randomly altered.}

\item{gender}{Character. Defaults to NULL, and is replaced.}

\item{birthyear}{Numeric. Defaults to NULL, and is replaced}

\item{tx_focus}{Character. Defaults to NULL, and is replaced}

\item{in_or_out}{Character. Defaults to NULL, and is replaced}

\item{...}{Additional parameters}
}
\value{

}
\description{
Generate a set of random patient variables of arbitrary length
}
\examples{
random_person_generator(100)
}
