% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{random_tx_generator}
\alias{random_tx_generator}
\title{Generate an entire treatment course, possibly given several specific parameters}
\usage{
random_tx_generator(
  num_obs,
  num_pts = length(num_obs),
  first_date = rep(as.Date("2017-11-13 11:30:51 UTC"), num_pts),
  identifiers = 1:num_pts,
  tx_days = NULL,
  ...
)
}
\arguments{
\item{num_obs}{Vector. Number of observations per person. Can be a vector.}

\item{num_pts}{Integer. Number of patients.}

\item{first_date}{Vector of dates with length = num_pts.}

\item{identifiers}{Vector. Optional names for individuals.}

\item{tx_days}{Numeric. Defaults to NULL and is rewritten. Defines typical length of treatment.}

\item{...}{Additional parameters.}
}
\value{
A data frame probably
}
\description{
Generate an entire treatment course, possibly given several specific parameters
}
\examples{
random_tx_generator(4)
random_tx_generator(num_obs = c(3, 5))
random_tx_generator(num_obs = 5, num_pts = 3)
random_tx_generator(1:3, first_date = c(as.Date("2017-11-20"),
                                      as.Date("2017-03-20"),
                                      as.Date("2017-12-20")))
random_tx_generator(num_obs = c(1, 4, 2),
                    first_date = c(as.Date("2017-11-20"),
                                        as.Date("2017-03-20"),
                                        as.Date("2017-12-20")),
                    num_pts = 3)
}
