% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funcs.R
\name{read.csv2_nf3}
\alias{read.csv2_nf3}
\title{Read external NF3 data}
\usage{
read.csv2_nf3(file, ...)
}
\arguments{
\item{file}{Location of the file}

\item{...}{Additional arguments}
}
\value{
a data frame
}
\description{
Read external NF3 data
}
\examples{
example_file <- tempfile(fileext = ".csv")
writeLines(c("Q1;Q2", "1,2"), example_file)
read.csv2_nf3(example_file)
}
