% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funcs.R
\name{read_excel_nf3}
\alias{read_excel_nf3}
\title{Read external NF3 data}
\usage{
read_excel_nf3(file, ...)
}
\arguments{
\item{file}{Location of the file}

\item{...}{Additional arguments}
}
\value{
a data.frame
}
\description{
Read external NF3 data
}
\examples{
\dontrun{
read_excel_nf3("NF3_export.xlsx")
}
}
