% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R
\name{scale_analysis}
\alias{scale_analysis}
\title{Summary analysis of scales using IRT}
\usage{
scale_analysis(
  scale.name,
  item.names,
  data,
  print.now = FALSE,
  irt.package = "ltm"
)
}
\arguments{
\item{scale.name}{String. The name of the scale of interest,
e.g., \code{"Suicide Risk"}.}

\item{item.names}{Character vector. The names of the items that comprise the scale.}

\item{data}{Data frame. The data set in which to find items for analysis.}

\item{print.now}{Logical. Should the item curves print on running, or
be stored for later use ONLY? Defaults to \code{FALSE}, which will not print on run.}

\item{irt.package}{Which underlying package to use for estimation of GRM? Defaults
to \code{'ltm'}, for backwards compatibility. May also be \code{'mirt'}.}
}
\value{
a list with the following named values:
\item{\code{new_data}}{The data set used for analysis}
\item{\code{histogram}}{Item response histogram}
\item{\code{min}}{Observed minimum item response mean}
\item{\code{max}}{Observed maximum item response mean}
\item{\code{zmin}}{Standardized obsered minimum item response mean}
\item{\code{zmax}}{Standardized obsered maximum item response mean}
\item{\code{floor}}{Percent of individuals at absolute minimum}
\item{\code{ceiling}}{Percent of individuals at absolute maximum}
\item{\code{grm}}{An object of type \code{grm} or \code{mirt} for scale analysis}
\item{\code{ICC}}{Item Characteristic Curves. One per item, faceted if \code{irt.package = 'mirt'}.
Either a plot object or the plot information, depending on the value of \code{print.now}.}
\item{\code{TIF}}{Either a plot object or the plot information, depending on the value of \code{print.now}.}
\item{\code{IIC}}{Either a plot object or the plot information, depending on the value of \code{print.now}.}
\item{\code{printed}}{Logical; were the IRT plots printed on run?}
\item{\code{info}}{Relative marginal information contribution table. Currently not working
for \code{irt.package = 'mirt'} (2020 AUG 25).}
\item{\code{tables}}{A list of tables of item response frequencies, of length k items}
\item{\code{cor}}{A polychoric correlation table of item responses}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{scale_analysis} is a summary function, especially useful in report writing.
Given some limited input, it returns a list of output objects which can
later be printed in reports or extracted in various ways. Use
\code{\link[=scale_analysis2]{scale_analysis2()}} instead.
}
\note{
This function will give an error if the \code{grm()} call doesn't work, which may
depend on having all observed levels of all items. Use \code{irt.package = 'mirt'}
in that case.
}
\examples{
\dontrun{
data(synthetic_data, package = "NorseResearch")
internal.out <- scale_analysis("Internal Avoidance", internal.names,
                               data = synthetic_data)
}


}
