% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scoreNames.nf3}
\alias{scoreNames.nf3}
\title{Lookup table for Exported scores to useful scales}
\format{
A data frame with 47 rows and 2 variables:
\describe{
\item{\code{ScoreName}}{character. Simplified name of the SCORE variable}
\item{\code{ScaleName}}{character. Nicely formatted Scale name}
}
}
\source{
Based on exports of data provided by NF, and more proximally \code{data-raw/scoring_objects_raw.R}.
}
\usage{
scoreNames.nf3
}
\description{
A linking table for SCORE variables and human-readable scale names
}
\details{
Could be incorporated into a scoring function.
}
\keyword{datasets}
