% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R
\name{summary.scale_analysis}
\alias{summary.scale_analysis}
\title{Summarizing scale_analysis objects}
\usage{
\method{summary}{scale_analysis}(object, ...)
}
\arguments{
\item{object}{Fitted object.}

\item{...}{Additional arguments passed to the summary method.}
}
\description{
\code{summary.scale_analysis} takes an object from \code{scale_analysis()} and
provides several summaries, but no plots.
}
