% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{summary_norms_MH_out}
\alias{summary_norms_MH_out}
\title{Summary norm data for mental health specialty clinics, NF2, Helse Førde data, from 2019.
Only includes mean, SD, and N.}
\format{
a table.
\describe{
\item{\code{var}}{Variable in each row. Values: Mean, SD, and N}
\item{\code{cog}}{Cognitive problems}
\item{\code{ALLOTHERS}}{Other scales in NF2.1}
}
}
\usage{
summary_norms_MH_out
}
\description{
Summary norm data for mental health specialty clinics, NF2, Helse Førde data, from 2019.
Only includes mean, SD, and N.
}
\keyword{datasets}
