% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuncs.R
\name{trigger_plot}
\alias{trigger_plot}
\title{Plotting trigger item predicted distributions.
NEEDS options to allow mirt objects with name,
picking from list of SA objects, or new computation from data.
Also to assign a new trigger item. Right now it does the first item.}
\usage{
trigger_plot(scaleName, mirt_obj, lim = c(-3, 3))
}
\arguments{
\item{scaleName}{A quoted simple scale name, e.g., \code{"somAnx"}.
Suitable for use with \code{scale.names}.}

\item{mirt_obj}{A fitted \code{mirt} \code{SingleGroupClass} object.}

\item{lim}{Interval for plotting, as a vector of two values.}
}
\value{
ggplot object, printed by default.
}
\description{
Plotting trigger item predicted distributions.
NEEDS options to allow mirt objects with name,
picking from list of SA objects, or new computation from data.
Also to assign a new trigger item. Right now it does the first item.
}
