% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rev.R
\name{check_rev}
\alias{check_rev}
\title{check if NORSE scores are reversed}
\usage{
check_rev(.data, verbose = FALSE, version = "NF2")
}
\arguments{
\item{.data}{The data to clean.}

\item{verbose}{Logical. Default is FALSE. If TRUE, will
return the list of correlations tested with their values.}

\item{version}{Which version of NF is being used. Acceptable values
are "NF2" and "NF3".}
}
\value{
logical. Is data likely reversed? TRUE if yes.
If \code{verbose} is \code{TRUE}, returns a named list.
}
\description{
check if NORSE scores are reversed
}
\examples{
data(synthetic_data, package = "NorseResearch")
check_rev(synthetic_data)
check_rev(synthetic_data, verbose = TRUE)

}
