% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{check_version_nf}
\alias{check_version_nf}
\title{Check which version(s) of the NF are present in a data set.}
\usage{
check_version_nf(dat)
}
\arguments{
\item{dat}{a data.frame}
}
\value{
a character vector containing one or more of the following: "2", "3".
If no matching items are found, will return "No NF items found."
}
\description{
Should not require the data to already have clean names. But needs to work when that is the case.
}
\examples{
/dontrun{
check_version_nf(dataGoesHere)
}
}
