% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_NF_names.R
\name{clean_NF_names}
\alias{clean_NF_names}
\title{Clean variable names of a Norse Feedback data frame}
\source{
https://drdoane.com/clean-consistent-column-names/,
with some edits by AAM.
}
\usage{
clean_NF_names(.data, unique = FALSE)
}
\arguments{
\item{.data}{The data to clean}

\item{unique}{Logical. Should unique variable names be permitted?}
}
\value{
A data frame with clean names
}
\description{
Clean variable names of a Norse Feedback data frame
}
\examples{
 clean_NF_names(
   c(
   "  a", "a  ",
   "a \%", "a", "$a", "$$$a", "GDP ($)", "GDP (us$)",
   "a (#)", "a & b", "#", "$",
   "a_cnt",
   "Aa&Bb", "camelCasePhrases",
   "AlphaBetaGamma", "Alpha       Beta", "Beta  !!! Gamma",
   "a + b", "a - b", "a * b"
   )
 )

 library(dplyr)
 df <- read.csv("data-raw/source_data") \%>\%
   clean_NF_names()

}
