% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingFuncs.R
\name{get_item_text}
\alias{get_item_text}
\title{Retrieving the (English) item text from an item number}
\usage{
get_item_text(target, table = item_descriptions)
}
\arguments{
\item{target}{the character string to look up, e.g., "Q26".}

\item{table}{the lookup table, defaults to \code{item_descriptions}.
Can be expanded to provide Norwegian or other lookup tables.}
}
\value{
a character string with the English-language item text.
}
\description{
\code{get_item_text} simply returns the English text, given a
character item number in the format "Q11".
}
