% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HF_research_data_2021}
\alias{HF_research_data_2021}
\title{Research-friendly data from Helse Førde 2021.}
\format{
A data frame with 18620 observations of 170 variables.
}
\source{
Processed from NORSEpkg::hf.all.scored.2021.05.10.
See \code{data-raw/HF_research_data_2021.R}.
}
\usage{
HF_research_data_2021
}
\description{
Anonymous research data from Helse Førde area (all clinics)
from May 2021, scored, with additional variables for external use.
}
\details{
Any values that were 99 are now NA, and these represent patient non-responses.

THIS DATA IS NOT PROVIDED as part of the package and must be received separately
for confidentiality. Once received, the .rda file can be saved in the
global environment as \code{HF_research_data_2021 <- load("HF_research_data_2021.rda")}.
This process means that it is not really accessible as part of the package,
but must be stored locally.

The codebook for the relevant items is provided here though.
}
\keyword{datasets}
