% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuncs.R
\name{item_plot}
\alias{item_plot}
\title{Plot histogram of NORSE items.}
\usage{
item_plot(item, ..., data = data.joined)
}
\arguments{
\item{item}{The name of the item to plot.}

\item{data}{A dataframe.}
}
\value{
Not much, just a plot.
}
\description{
\code{item_plot} plots histogram of item responses from a dataframe.
}
\examples{
item_plot("Q71", data.joined)
item_plot(Q71, data.joined)  # error

# Can be done in lapply:
needs.names <- c("Q71", "Q72", "Q74", "Q152", "Q153")
lapply(needs.names, item_plot)
}
