% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_score.R
\name{nf_score}
\alias{nf_score}
\title{Scoring mixed NF data}
\usage{
nf_score(
  dat,
  versions_to_score = c("2", "3"),
  version_variable = "Ver_10",
  process_vars = TRUE,
  raw_scores = TRUE,
  factor_scores = FALSE
)
}
\arguments{
\item{dat}{A data.frame.}

\item{versions_to_score}{Character vector. Which versions should be scored
if present. defaults to \code{c("2", "3")}, which scores the 2 and 3. Only 2 and 3
are available now. Ideally will add versions over time.}

\item{version_variable}{Character. Name of the variable in \code{dat} that contains version information.}

\item{process_vars}{Logical. Should process variables (Alliance, medications) be included? Defaults to \code{TRUE}.}

\item{raw_scores}{Logical. Should raw scale scores be computed? Defaults to \code{TRUE}.}

\item{factor_scores}{Logical. Should factor-scored scale scores be computed? Defaults to \code{FALSE}.}
}
\value{
A data.frame/tibble
}
\description{
This function should take a dataset and add new variables as needed.
It is obsolete and should be replaced with \code{score_all()}.
}
\examples{
/dontrun{
nf_score()
}
}
