% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaningFuncs.R
\name{nicer_id_var}
\alias{nicer_id_var}
\title{create a better-named patient_id variable in NF-like data}
\usage{
nicer_id_var(.data, id = `respondent id`, keep_old_vars = TRUE)
}
\arguments{
\item{.data}{Data frame}

\item{id}{Existing patient id variable name. Default is \code{`respondent id`},
a particularly terrible instance.}

\item{keep_old_vars}{Logical. Should the returned dataset include the
existing id variable? Defaults to \code{TRUE}. Implemented using \code{dplyr::mutate()}
\code{.keep} option.}
}
\value{
A data frame or tibble.
}
\description{
create a better-named patient_id variable in NF-like data
}
\examples{
nicer_id_var(hf.scored.2019)
}
