% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorseResearch-package.R
\docType{package}
\name{NorseResearch-package}
\alias{NorseResearch}
\alias{NorseResearch-package}
\title{NorseResearch: Utilities for analyzing Norse Feedback data}
\description{
Provides basic functions and data information (not data itself) for research purposes.
}
\author{
\strong{Maintainer}: Andrew Mcaleavey \email{andrew.athan.mcaleavey@helse-forde.no} (\href{https://orcid.org/0000-0001-5986-2033}{ORCID})

Authors:
\itemize{
  \item Andrew Mcaleavey \email{andrew.athan.mcaleavey@helse-forde.no} (\href{https://orcid.org/0000-0001-5986-2033}{ORCID})
}

}
\keyword{internal}
