% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R
\name{plot.scale_analysis}
\alias{plot.scale_analysis}
\title{Plotting scale_analysis objects}
\usage{
\method{plot}{scale_analysis}(x)
}
\arguments{
\item{x}{fitted scale analysis object (named list).}

\item{lim}{limits for x-axis. Defaults to c(-4, 4).}
}
\description{
\code{plot.scale_analysis} takes an object from \code{scale_analysis()} and
provides several printed plots, for fast visual analysis.
}
