% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_data.R
\name{random_norse_data}
\alias{random_norse_data}
\title{Create a synthetic data set of NF-like data}
\usage{
random_norse_data(num_ppl, ...)
}
\arguments{
\item{num_ppl}{Numeric. Number of patients.}

\item{...}{Additional parameters}
}
\value{
A tibble.
}
\description{
Create a synthetic data set of NF-like data
}
\examples{
random_norse_data(2, num_obs = c(3, 2))
random_norse_data(2, num_obs = 5)
random_norse_data(2)
}
\seealso{
\code{\link{random_assessment_generator}}, \code{\link{random_person_generator}},
\code{\link{random_tx_generator}}
}
