% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_funcs.R
\name{read.csv_nf3}
\alias{read.csv_nf3}
\title{Read external NF3 data}
\usage{
read.csv_nf3(file, ...)
}
\arguments{
\item{file}{Location of the file}

\item{...}{Additional arguments}
}
\value{
a data.frame
}
\description{
Read external NF3 data
}
\examples{
read.csv_nf3("HF_research_data.csv")
}
