% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R
\name{replot.grm}
\alias{replot.grm}
\title{Plotting ICCs from grm()-derived plot information}
\usage{
replot.grm(fit, type = "ICC")
}
\description{
\code{replot.grm} takes plot information derived from
\code{ltm::plot.grm(plot = FALSE)}, and returns a \code{ggplot}-based series of plots
}
