% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{rev_score}
\alias{rev_score}
\title{Reverse score NORSE items (1-7).}
\usage{
rev_score(x)
}
\arguments{
\item{x}{A vector of item responses.}
}
\value{
A reverse-coded vector of \code{x}.
}
\description{
\code{rev_score} takes a vector and returns its 1-7 revers
}
\examples{
data <- data.frame(Q27 = c(1, 2, 2, 1), Q28 = c(7, 6, 6, 7))
dplyr::mutate(data, Q27 = rev_score(Q27))
}
