% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R, R/reportingFuncs.R
\name{scale_analysis}
\alias{scale_analysis}
\title{Summary analysis of scales using IRT}
\usage{
scale_analysis(
  scale.name,
  item.names,
  data,
  IRTpackage = "ltm",
  print.now = FALSE
)

scale_analysis(
  scale.name,
  item.names,
  data,
  IRTpackage = "ltm",
  print.now = FALSE
)
}
\arguments{
\item{scale.name}{A character string specifying the name of the scale for labeling.}

\item{item.names}{A character vector of item names (column names in \code{data}).}

\item{data}{A data frame containing the item responses.}

\item{IRTpackage}{A character string specifying which IRT package to use.
Options: \code{"ltm"} (default) or \code{"mirt"}.}

\item{print.now}{Logical. If \code{TRUE}, plots for item characteristic curves (ICC) and
test information function (TIF) are displayed immediately. Default is \code{FALSE}.}

\item{irt.package}{Which underlying package to use for estimation of GRM? Defaults
to \code{'ltm'}, for backwards compatibility. May also be \code{'mirt'}.}
}
\value{
a list with the following named values:
\itemize{
\item{\code{new_data}}{: The data set used for analysis}
\item{\code{histogram}} {Item response histogram}
\item{\code{min}} {Observed minimum item response mean}
\item{\code{max}} {Observed maximum item response mean}
\item{\code{zmin}} {Standardized obsered minimum item response mean}
\item{\code{zmax}} {Standardized obsered maximum item response mean}
\item{\code{floor}} {Percent of individuals at absolute minimum}
\item{\code{ceiling}} {Percent of individuals at absolute maximum}
\item{\code{grm}} {An object of type \code{grm} or \code{mirt} for scale analysis}
\item{\code{ICC}} {Item Characteristic Curves. One per item, faceted if \code{irt.package = 'mirt'}.
Either a plot object or the plot information,
depending on the value of \code{print.now}}
\item{\code{TIF}} {Either a plot object or the plot information,
depending on the value of \code{print.now}}
\item{\code{IIC}} {Either a plot object or the plot information,
depending on the value of \code{print.now}}
\item{\code{printed}} {logial; Were the IRT plots printed on run?}
\item{\code{info}} {Relative marginal information contribution table. Currently not working
for \code{irt.package = 'mirt'} (2020 AUG 25).}
\item{\code{tables}} {A list of tables if item response frequency, of length k items}
\item{\code{cor}} {A polychoric correlation table of item responses}
}

A list containing:
\item{new_data}{Subset of \code{data} containing only the selected items.}
\item{histogram}{A ggplot histogram of the scale score distribution.}
\item{min, max}{Observed minimum and maximum scale scores.}
\item{zmin, zmax}{Standardized minimum and maximum scores.}
\item{floor, ceiling}{Proportion of responses at the minimum and maximum score.}
\item{alpha}{Cronbach's alpha and other reliability statistics.}
\item{grm}{The fitted graded response model (\code{ltm::grm} or \code{mirt::mirt}).}
\item{ICC, TIF, IIC}{Item characteristic curves, test information, and item information.}
\item{printed}{Logical indicating if plots were printed (\code{TRUE}/\code{FALSE}).}
\item{info}{Item information statistics from the IRT model.}
\item{tables}{Frequency tables for each item.}
\item{cor}{Polychoric correlation matrix of the items.}
}
\description{
\code{scale_analysis} is a summary function, especially useful in report writing.
Given some limited input, it returns a list of output objects which can
later be printed in reports or extracted in various ways.

This function performs scale analysis by computing descriptive statistics, reliability,
and item response theory (IRT) models using either the \code{{ltm}} or \code{{mirt}} package.
}
\note{
This function will give an error if the \code{grm()} call doesn't work, which may
depend on having all observed levels of all items. Use \code{irt.package = 'mirt'}
in that case.
}
\examples{
internal.out <- scale_analysis("Internal Avoidance", internal.names, data = data2017_18)


\dontrun{
df <- data.frame(
Q1 = sample(1:5, 100, replace = TRUE),
Q2 = sample(1:5, 100, replace = TRUE),
Q3 = sample(1:5, 100, replace = TRUE)
)
result <- scale_analysis("Example Scale", c("Q1", "Q2", "Q3"),
                         df, IRTpackage = "mirt")
print(result$histogram)
}
}
