% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFuncs.R
\name{scale_plot}
\alias{scale_plot}
\title{Plot an annotated histogram for NF scales}
\usage{
scale_plot(data, item.names, title_obj)
}
\arguments{
\item{data}{The data set on which to compute.}

\item{item.names}{The names of items to include, in a character vector.}

\item{title_obj}{Title of the plot.}
}
\value{
A ggplot object
}
\description{
Plot an annotated histogram for NF scales
}
\examples{
scale_plot(hb.scored.2019, cog.names, "Cognitive Problems")
}
