% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{score_all}
\alias{score_all}
\title{Score all versions of the NF3 in raw scale form}
\usage{
score_all(
  dat,
  process_vars = TRUE,
  versions = c("2", "3"),
  version_variable = "Ver_10"
)
}
\arguments{
\item{dat}{A dataset. Should have clean names that have been simplified to include
only item indicators. For example, \code{M53_Q142_1} should instead be \code{Q142}.}

\item{process_vars}{Logical. Should the process variables be included?}

\item{versions}{Which versions of the NF do you want to score? Includes both
2.x and 3.x by default.}

\item{version_variable}{Character name of the variable encoding the version
to use for each row.}
}
\value{
A data.frame/tibble with new variables. The original item responses remain unchanged.
}
\description{
This function AUTOMATICALLY treats data with -98 and -99 values. The assumption is that
-98 means "not relevant to me" and -99 means "Prefer not to answer". For all scores,
-99s are considered NA when calculating scores. For all scales EXCEPT PROCESS VARIABLES -98s
mean 1, the lowest possible severity. For Alliance and Preference scales, -98 means NA.
}
\examples{
\dontrun{
score_all(dataGoesHere)
}
}
