% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{score_all}
\alias{score_all}
\title{Score all versions of the NF3 in raw scale form}
\usage{
score_all(
  dat,
  process_vars = TRUE,
  versions = c("2", "3"),
  version_variable = "Ver_10"
)
}
\arguments{
\item{dat}{A dataset. Should have clean names that have been simplified to include
only item indicators. For example, \code{M53_Q142_1} should instead by \code{Q142}.}

\item{process_vars}{logical. Should the process variables be included?}

\item{versions}{Which versions of the NF do you want to score? Includes both
2.x and 3.x by default.}

\item{version_variable}{Character name of the variable encoding the version
to use for each row.}
}
\value{
A data.frame/tibble with new variables.
}
\description{
Score all versions of the NF3 in raw scale form
}
\examples{
/dontrun{
score_all(dataGoesHere)
}
}
