% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{score_all_nf3}
\alias{score_all_nf3}
\title{Score the NF3 scales}
\usage{
score_all_nf3(dat, process_vars = TRUE)
}
\arguments{
\item{dat}{Data}

\item{process_vars}{logical. Should the process variables be included?}
}
\value{
A data frame with additional variables
}
\description{
Score the NF3 scales
}
\examples{
/dontrun{
score_all_nf3(HF_research_data_2021)
}
}
