% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{score_normed_NF}
\alias{score_normed_NF}
\title{Norm NF data}
\usage{
score_normed_NF(
  dat,
  scale,
  normTable = summary_norms_MH_out,
  inputFormat = "raw",
  outputFormat = "Z"
)
}
\arguments{
\item{dat}{A data.frame. Data set on which to add normed variables.}

\item{scale}{String. Which scale should be normed. Does one scale at a time.}

\item{normTable}{Which norm table should be used. Defaults to the 2019
MH outpatient first visit norms ("summary_norms_MH_out").
A custom normTable can be supplied, or
several options within the package could be developed.}

\item{inputFormat}{What is the scale/format of the data to be normed?
Defaults to "raw", which indicates scale scores will range from 1-7.
Alternatives may include previously normed data on Z, T, or other scales
if in use.}

\item{outputFormat}{What is the intended normalized scale? Defaults
to "Z", indicating mean = 0 and SD = 1. Alternatives could include T or
other}
}
\value{
A tibble. the input data.frame with additional variable(s) for
normed scores. Normed scale score variables take the name format
scaleName_normed_normTable_outputFormat, such as "cog_normed_MHout_Z".
}
\description{
Norm NF data
}
\examples{
data(synthetic_data, package = "NorseResearch")
testout <- score_normed_NF(synthetic_data, scale = "cog")
}
