% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoringFuncs.R
\name{score_NORSE_overunder}
\alias{score_NORSE_overunder}
\title{Scoring Over-Under style}
\usage{
score_NORSE_overunder(
  dat,
  vars = names(dat),
  trigger = vars[1],
  closing.thresh = 0,
  opening.thresh = 0
)
}
\arguments{
\item{dat}{a data_frame.}

\item{vars}{the variables to score, defaults to all variables in \code{dat}.}

\item{trigger}{the trigger item, defaults to the first item in \code{dat}.}

\item{closing.thresh}{the closing threshold value. Default is 0.}

\item{opening.thresh}{the opening threshold value. Default is 0.}
}
\value{
a vector of new subscale scores
}
\description{
\code{score_NORSE_overunder} is a general method for scoring
over-under closing and opening thresholds.
}
\details{
The principle of this scoring is that WHILE THE SCALE IS OPEN,
scores should be centered on the closing threshold, and
WHILE THE SCALE IS CLOSED, scores should be centered on the
opening threshold.

Pseudocode scoring:
IF ONLY the trigger is completed
Score is the trigger score MINUS the opening threshold
ELSE
Score is the mean score as computed standardly,
MINUS the closing threshold

This will mean that there will be some NEGATIVE scores for open scales
}
\examples{
test_data <- data.frame(Q1 = c(1, 2, 3, 5, 3, 2, NA, NA),
                        Q2 = c(3, 4, 2, 5, 2, NA, 1, NA),
                        Q3 = c(4, 3, 5, 2, NA, NA, NA, NA))

score_NORSE_overunder(dat = test_data, vars = c("Q1", "Q2", "Q3"))

# The same calculation can be used inside a dplyr::mutate() call.
dplyr::mutate(test_data, subscale_score = score_NORSE_trigger(test_data))
dplyr::mutate(test_data, subscale_score = score_NORSE_trigger(test_data, trigger = "Q3"))
}
