% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R
\name{summary.scale_analysis}
\alias{summary.scale_analysis}
\title{Summarizing scale_analysis objects}
\usage{
\method{summary}{scale_analysis}(x)
}
\arguments{
\item{x}{fitted object.}

\item{lim}{limits for x-axis. Defaults to c(-4, 4).}
}
\description{
\code{summary.scale_analysis} takes an object from \code{scale_analysis()} and
provides several summaries, but no plots.
}
