% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_data}
\alias{synthetic_data}
\title{Synthetic Data (N = 500)}
\format{
A data frame with 2883 observations of 132 variables.
\describe{
\itemize{
\item{\code{anon_id}} {Numeric. Unique patient number within this data only, anonymized.}
\item{\code{date}} {Date. Date (and time) NF was completed.}
\item{\code{pt_first_date}} {Date. First date in the data for each \code{respondent_id}.}
\item{\code{pt_total_obs}} {Numeric. Total number of observations (rows) of data for this patient.}
\item{\code{tx_focus}} {Character. What was the focus of \code{treatment_name} for this observation.
Values are "Sub", "MH", or "Unclear".}
\item{\code{in_or_out}} {Character. Is the current \code{treatment_name} inpatient, outpatient, or unclear.}
\item{\code{birthyear}} {Character. Year of birth for each \code{respondent_id}.}
\item{\code{gender}} {Character. Options are "male" and "female".}
\item{code{OTHERS}} {Others including scale and item names.}
}
}
}
\source{
See \code{data-raw/synthetic_data.R}.
}
\usage{
synthetic_data
}
\description{
This is not real data, it is essentially random numbers in the form of data
very similar to what Norse Feedback looks like. it is based on the function
random_norse_data().
}
\keyword{datasets}
