% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRTFuncs.R
\name{tif.grm}
\alias{tif.grm}
\title{Plotting Test Information Functions in {ltm}}
\usage{
tif.grm(x, lim = c(-4, 4))
}
\arguments{
\item{x}{fitted object.}

\item{lim}{limits for x-axis. Defaults to c(-4, 4).}
}
\description{
\code{tif.grm} takes a fitted ltm model and plots its Test Information Function
}
